modify the way of deploing openpitrix with k8s

+++ test/init/init_config.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
 
-mkdir -p ~/.openpitrix/
-cp config.json ~/.openpitrix/
+if [ ! -f "~/.openpitrix/config.json" ];then
+  mkdir -p ~/.openpitrix/
+  cp config.json ~/.openpitrix/
+fi
